// visualization/src/generated_types/BorrowPcgActionKindDebugRepr.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BorrowPcgActionKindDiscriminants } from "./BorrowPcgActionKindDiscriminants";

export type BorrowPcgActionKindDebugRepr = { type: BorrowPcgActionKindDiscriminants, data: string, };