// visualization/src/generated_types/RepackOp.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CapabilityKind } from "./CapabilityKind";
import type { RegainedCapability } from "./RegainedCapability";
import type { RepackCollapse } from "./RepackCollapse";
import type { RepackExpand } from "./RepackExpand";
import type { Weaken } from "./Weaken";

export type RepackOp = { "type": "StorageDead", "data": string } | { "type": "IgnoreStorageDead", "data": string } | { "type": "Weaken", "data": Weaken<string, CapabilityKind, CapabilityKind> } | { "type": "Expand", "data": RepackExpand } | { "type": "Collapse", "data": RepackCollapse<string> } | { "type": "DerefShallowInit", "data": [string, string] } | { "type": "RegainLoanedCapability", "data": RegainedCapability<string> };